backend/provider-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.6.0
redis==5.0.1
tenacity==8.2.3
boto3==1.34.0
botocore==1.34.0
asyncpg==0.29.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
python-json-logger==2.0.7
hypothesis==6.98.0
aiokafka==0.10.0
circuitbreaker==2.0.0
starlette==0.36.3
hypothesis==6.98.0
pytest-asyncio==0.23.6
